Refactor controller methods, non-breaking change
