+++ server/src/controllers/asset-media.controller.ts
@@ -1,10 +1,12 @@
 import {
   Body,
   Controller,
+  HttpCode,
   HttpStatus,
   Inject,
   Param,
   ParseFilePipe,
+  Post,
   Put,
   Res,
   UploadedFiles,
@@ -15,13 +17,19 @@ import { Response } from 'express';
 import { EndpointLifecycle } from 'src/decorators';
 import { AssetMediaResponseDto, AssetMediaStatusEnum } from 'src/dtos/asset-media-response.dto';
 import { AssetMediaReplaceDto, UploadFieldName } from 'src/dtos/asset-media.dto';
+import {
+  AssetBulkUploadCheckResponseDto, CheckExistingAssetsResponseDto
+} from 'src/dtos/asset-media-response.dto';
+import {
+  AssetBulkUploadCheckDto,
+  CheckExistingAssetsDto
+} from 'src/dtos/asset-media.dto';
 import { AuthDto } from 'src/dtos/auth.dto';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { Auth, Authenticated } from 'src/middleware/auth.guard';
 import { FileUploadInterceptor, Route, UploadFiles, getFiles } from 'src/middleware/file-upload.interceptor';
 import { AssetMediaService } from 'src/services/asset-media.service';
 import { FileNotEmptyValidator, UUIDParamDto } from 'src/validation';
-
 @ApiTags('Asset')
 @Controller(Route.ASSET)
 export class AssetMediaController {
@@ -52,5 +60,31 @@ export class AssetMediaController {
       res.status(HttpStatus.OK);
     }
     return responseDto;
+  }
+
+  /**
+   * Checks if multiple assets exist on the server and returns all existing - used by background backup
+   */
+  @Post('/exist')
+  @HttpCode(HttpStatus.OK)
+  @Authenticated()
+  checkExistingAssets(
+    @Auth() auth: AuthDto,
+    @Body() dto: CheckExistingAssetsDto,
+  ): Promise<CheckExistingAssetsResponseDto> {
+    return this.service.checkExistingAssets(auth, dto);
+  }
+
+  /**
+   * Checks if assets exist by checksums
+   */
+  @Post('/bulk-upload-check')
+  @HttpCode(HttpStatus.OK)
+  @Authenticated()
+  checkBulkUpload(
+    @Auth() auth: AuthDto,
+    @Body() dto: AssetBulkUploadCheckDto,
+  ): Promise<AssetBulkUploadCheckResponseDto> {
+    return this.service.bulkUploadCheck(auth, dto);
   }
 }
