+++ server/src/controllers/asset-v1.controller.ts
@@ -2,7 +2,6 @@ import {
   Body,
   Controller,
   Get,
-  HttpCode,
   HttpStatus,
   Inject,
   Next,
@@ -16,20 +15,8 @@ import {
 } from '@nestjs/common';
 import { ApiBody, ApiConsumes, ApiHeader, ApiTags } from '@nestjs/swagger';
 import { NextFunction, Response } from 'express';
-import { AssetResponseDto } from 'src/dtos/asset-response.dto';
-import {
-  AssetBulkUploadCheckResponseDto,
-  AssetFileUploadResponseDto,
-  CheckExistingAssetsResponseDto,
-} from 'src/dtos/asset-v1-response.dto';
-import {
-  AssetBulkUploadCheckDto,
-  AssetSearchDto,
-  CheckExistingAssetsDto,
-  CreateAssetDto,
-  GetAssetThumbnailDto,
-  ServeFileDto,
-} from 'src/dtos/asset-v1.dto';
+import { AssetFileUploadResponseDto } from 'src/dtos/asset-v1-response.dto';
+import { CreateAssetDto, GetAssetThumbnailDto, ServeFileDto } from 'src/dtos/asset-v1.dto';
 import { AuthDto, ImmichHeader } from 'src/dtos/auth.dto';
 import { ILoggerRepository } from 'src/interfaces/logger.interface';
 import { AssetUploadInterceptor } from 'src/middleware/asset-upload.interceptor';
@@ -108,46 +95,5 @@ export class AssetControllerV1 {
     @Query() dto: GetAssetThumbnailDto,
   ) {
     await sendFile(res, next, () => this.service.serveThumbnail(auth, id, dto), this.logger);
-  }
-
-  /**
-   * Get all AssetEntity belong to the user
-   */
-  @Get('/')
-  @ApiHeader({
-    name: 'if-none-match',
-    description: 'ETag of data already cached on the client',
-    required: false,
-    schema: { type: 'string' },
-  })
-  @Authenticated()
-  getAllAssets(@Auth() auth: AuthDto, @Query() dto: AssetSearchDto): Promise<AssetResponseDto[]> {
-    return this.service.getAllAssets(auth, dto);
-  }
-
-  /**
-   * Checks if multiple assets exist on the server and returns all existing - used by background backup
-   */
-  @Post('/exist')
-  @HttpCode(HttpStatus.OK)
-  @Authenticated()
-  checkExistingAssets(
-    @Auth() auth: AuthDto,
-    @Body() dto: CheckExistingAssetsDto,
-  ): Promise<CheckExistingAssetsResponseDto> {
-    return this.service.checkExistingAssets(auth, dto);
-  }
-
-  /**
-   * Checks if assets exist by checksums
-   */
-  @Post('/bulk-upload-check')
-  @HttpCode(HttpStatus.OK)
-  @Authenticated()
-  checkBulkUpload(
-    @Auth() auth: AuthDto,
-    @Body() dto: AssetBulkUploadCheckDto,
-  ): Promise<AssetBulkUploadCheckResponseDto> {
-    return this.service.bulkUploadCheck(auth, dto);
   }
 }
