+++ server/src/dtos/asset-media-response.dto.ts
@@ -9,3 +9,27 @@ export class AssetMediaResponseDto {
   status!: AssetMediaStatusEnum;
   id!: string;
 }
+
+export enum AssetUploadAction {
+  ACCEPT = 'accept',
+  REJECT = 'reject',
+}
+
+export enum AssetRejectReason {
+  DUPLICATE = 'duplicate',
+  UNSUPPORTED_FORMAT = 'unsupported-format',
+}
+
+export class AssetBulkUploadCheckResult {
+  id!: string;
+  action!: AssetUploadAction;
+  reason?: AssetRejectReason;
+  assetId?: string;
+}
+
+export class AssetBulkUploadCheckResponseDto {
+  results!: AssetBulkUploadCheckResult[];
+}
+export class CheckExistingAssetsResponseDto {
+  existingIds!: string[];
+}
