+++ server/src/interfaces/asset.interface.ts
@@ -170,6 +170,7 @@ export interface IAssetRepository {
   getByChecksum(libraryId: string | null, checksum: Buffer): Promise<AssetEntity | null>;
   getUploadAssetIdByChecksum(ownerId: string, checksum: Buffer): Promise<string | undefined>;
   getByAlbumId(pagination: PaginationOptions, albumId: string): Paginated<AssetEntity>;
+  getByDeviceIds(ownerId: string, deviceId: string, deviceAssetIds: string[]): Promise<AssetEntity[]>;
   getByUserId(pagination: PaginationOptions, userId: string, options?: AssetSearchOptions): Paginated<AssetEntity>;
   getById(
     id: string,
