+++ server/src/repositories/asset-v1.repository.ts
@@ -1,6 +1,8 @@
 import { Injectable } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
-import { AssetSearchDto, CheckExistingAssetsDto } from 'src/dtos/asset-v1.dto';
+import { CheckExistingAssetsDto } from 'src/dtos/asset-media.dto';
+import { AssetSearchDto } from 'src/dtos/asset.dto';
+
 import { AssetEntity } from 'src/entities/asset.entity';
 import { AssetCheck, AssetOwnerCheck, IAssetRepositoryV1 } from 'src/interfaces/asset-v1.interface';
 import { OptionalBetween } from 'src/utils/database';
